fix - log_to_opentelemetry
